refactor: Profile 관련된 부분 px 을 vh 단위로 수정
