+++ src/components/ProfileMenuContainer/ProfileMenuContainer.js
@@ -9,18 +9,14 @@ import { useDispatch } from 'react-redux';
 import profileExample from '@assets/images/profile_example.png';
 import { setLogout } from '@store/Auth/auth';
 
-import Icon from '@components/Icon';
+import Icon from '@components/IconTemp';
 import ProfileIcon from '@components/ProfileIcon';
 
 const Wrapper = styled.div`
-  @media only screen and (max-width: 480px) {
-    display: none;
-  }
-
   position: absolute;
-  top: 53px;
-  right: 105px;
-  height: 73px;
+  top: 5.3vh;
+  right: 10.5vh;
+  height: 7.3vh;
   display: flex;
   flex-direction: row;
   align-items: center;
@@ -43,25 +39,25 @@ const WrapMenu = styled.div`
 
 const Name = styled.div`
   font-family: AppleSDGothicNeoEB00;
-  font-size: 20px;
-  padding-left: 29px;
-  padding-right: 10px;
+  font-size: 1.5vh;
+  padding-left: 2.9vh;
+  padding-right: 1vh;
   user-select: none;
   color: #3d3d3d;
 `;
 
 const List = styled.ul`
-  width: 143px;
-  padding: 18.5px 0px 18.5px 0px;
+  width: 14.3vh;
+  padding: 1.85vh 0vh 1.85vh 0vh;
   position: absolute;
-  top: 70px;
-  right: 0px;
+  top: 7vh;
+  right: 0;
   z-index: 101;
   background-color: #fff;
   transition: 0.25s ease all;
   transform: scale(0);
   transform-origin: 0 1;
-  border-radius: 10px;
+  border-radius: 1vh;
   box-shadow: 0 12px 24px 0 rgba(4, 4, 161, 0.15);
   display: flex;
   flex-direction: column;
@@ -72,9 +68,9 @@ const List = styled.ul`
 `;
 
 const Item = styled.li`
-  width: 87px;
-  padding-top: 12.5px;
-  padding-bottom: 12.5px;
+  width: 8.7vh;
+  padding-top: 1.25vh;
+  padding-bottom: 1.25vh;
   display: flex;
   align-items: center;
   justify-content: flex-start;
@@ -83,7 +79,7 @@ const Item = styled.li`
 const Each = styled.div`
   user-select: none;
   font-family: AppleSDGothicNeoM00;
-  font-size: 20px;
+  font-size: 1.5vh;
   color: #9e9e9e;
   &:hover {
     color: #f2886b;
@@ -114,7 +110,14 @@ export default function ProfileMenuContainer({ name, src }) {
             <Each onClick={() => history.push('/mypage')}>마이페이지</Each>
           </Item>
           <Item>
-            <Each onClick={() => dispatch(setLogout())}>로그아웃</Each>
+            <Each
+              onClick={() => {
+                dispatch(setLogout());
+                history.push('/');
+              }}
+            >
+              로그아웃
+            </Each>
           </Item>
         </List>
       </WrapMenu>
